Refactor: move device management into their own file
Part of the ongoing breakup of binding.py
